+++ include/pieces.hpp
@@ -8,7 +8,7 @@ class Piece {
 public:
   virtual ~Piece() {}
   virtual char getType() = 0;
-  virtual short getValue() = 0;
+  virtual float getValue() = 0;
   virtual bool isWhite() = 0;
   virtual int getRow() = 0;
   virtual int getCol() = 0;
@@ -21,10 +21,10 @@ class King : public Piece {
 public:
   ~King() {}
   King(bool w, int r, int c) :
-    type{'K'}, white{w}, value{0}, row{r}, col{c} {}
+    type{'K'}, white{w}, value{0.f}, row{r}, col{c} {}
 
   char getType() override { return type; }
-  short getValue() override { return value; }
+  float getValue() override { return value; }
   bool isWhite() override { return white; }
   int getRow() override { return row; }
   int getCol() override { return col; }
@@ -34,7 +34,7 @@ public:
 private:
   char type;
   bool white;
-  short value;
+  float value;
   int row;
   int col;
 };
@@ -44,10 +44,10 @@ class Queen : public Piece {
 public:
   ~Queen() {}
   Queen(bool w, int r, int c) :
-    type{'Q'}, white{w}, value{9}, row{r}, col{c} {}
+    type{'Q'}, white{w}, value{9.f}, row{r}, col{c} {}
 
   char getType() override { return type; }
-  short getValue() override { return value; }
+  float getValue() override { return value; }
   bool isWhite() override { return white; }
   int getRow() override { return row; }
   int getCol() override { return col; }
@@ -57,7 +57,7 @@ public:
 private:
   char type;
   bool white;
-  short value;
+  float value;
   int row;
   int col;
 };
@@ -67,10 +67,10 @@ class Rook : public Piece {
 public:
   ~Rook() {}
   Rook(bool w, int r, int c) :
-    type{'R'}, white{w}, value{5}, row{r}, col{c} {}
+    type{'R'}, white{w}, value{5.f}, row{r}, col{c} {}
 
   char getType() override { return type; }
-  short getValue() override { return value; }
+  float getValue() override { return value; }
   bool isWhite() override { return white; }
   int getRow() override { return row; }
   int getCol() override { return col; }
@@ -80,7 +80,7 @@ public:
 private:
   char type;
   bool white;
-  short value;
+  float value;
   int row;
   int col;
 };
@@ -90,10 +90,10 @@ class Bishop : public Piece {
 public:
   ~Bishop() {}
   Bishop(bool w, int r, int c) :
-    type{'B'}, white{w}, value{3}, row{r}, col{c} {}
+    type{'B'}, white{w}, value{3.f}, row{r}, col{c} {}
 
   char getType() override { return type; }
-  short getValue() override { return value; }
+  float getValue() override { return value; }
   bool isWhite() override { return white; }
   int getRow() override { return row; }
   int getCol() override { return col; }
@@ -103,7 +103,7 @@ public:
 private:
   char type;
   bool white;
-  short value;
+  float value;
   int row;
   int col;
 };
@@ -113,10 +113,10 @@ class Knight : public Piece {
 public:
   ~Knight() {}
   Knight(bool w, int r, int c) :
-    type{'N'}, white{w}, value{3}, row{r}, col{c} {}
+    type{'N'}, white{w}, value{3.f}, row{r}, col{c} {}
 
   char getType() override { return type; }
-  short getValue() override { return value; }
+  float getValue() override { return value; }
   bool isWhite() override { return white; }
   int getRow() override { return row; }
   int getCol() override { return col; }
@@ -126,7 +126,7 @@ public:
 private:
   char type;
   bool white;
-  short value;
+  float value;
   int row;
   int col;
 };
@@ -136,10 +136,10 @@ class Pawn : public Piece {
 public:
   ~Pawn() {}
   Pawn(bool w, int r, int c) :
-    type{'P'}, white{w}, value{1}, row{r}, col{c} {}
+    type{'P'}, white{w}, value{1.f}, row{r}, col{c} {}
 
   char getType() override { return type; }
-  short getValue() override { return value; }
+  float getValue() override { return value; }
   bool isWhite() override { return white; }
   int getRow() override { return row; }
   int getCol() override { return col; }
@@ -149,7 +149,7 @@ public:
 private:
   char type;
   bool white;
-  short value;
+  float value;
   int row;
   int col;
 };
